Explicitly ignore returns from Span functions (#7)
* Add -Wunmatched_returns, -Werror_handling, and -Wunderspecs flags for Dialyzer

As reported in https://github.com/appsignal/appsignal-elixir-phoenix/issues/5,
there are some warning we throw that we should include on CI.

* Explicitly ignore return from span functions in Appsignal.Phoenix.{Application, Channel, View}

Since the `Span.set_attribute/3` function takes and returns `Span`s and `nil`s,
we don't care about the return value here, as either case is accounted for.

The Application starts the EventHandler, which logs errors itself. Since we
don't do anyhthing else there, we don't need to catch the return.