Update Elixir and OTP versions in CI matrix (#51)
* Update Elixir and OTP versions in CI matrix

Instead of periodically updating the CI matrix file manually, this patch
adds a script to download the versions.rb file from the main
appsignal-elixir repository, then generate a CI matrix file from that.

That way, this repository's CI matrix always matches the main one,
without any work having to be done to keep them in sync.

* Depend on telemetry ~> 0.4 on OTP < 21

Telemetry 1.0 doesn't run on OTP 20, but that version is included in the
list of supported OTP versions. This patch depends on telemetry ~> 0.4
on OTP 20.

* Depend on phoenix_live_view < 0.17.4 on OTP < 21

phoenix_live_view >= 0.17.4 doesn't run on OTP 20, but versions before
that do. This patch depends on phoenix_live_view version >= 0.9 and <
0.17.4 on OTP 20.

[skip-changeset]