Add warning for using Phoenix.View on Phoenix 1.7
Phoenix 1.7 apps don't have a :view in their web modules. The closest
thing is :html, which is similar, but doesn't have Phoenix.View anymore,
which is what Appsignal.Phoenix.View depended on.

To make sure users don't move their use Appsignal.Phoenix.View line to
the html function in their Phoenix 1.7 web module, this patch makes sure
that whatever Appsignal.Phoenix.View is used in has a __templates__/0
and a render/2 function before doing anything. If not, it'll print a
warning:

> 10:20:44.354 [warning] AppSignal.Phoenix.View NOT attached to Elixir.AppsignalPhoenixExampleWeb.Layouts
>
> Template rendering instrumentation is now set up automatically, so using
> Appsignal.Phoenix.View in your app's web module is no longer needed.

This warning is noisy, as it prints for every view module in the app.

Also, it's printed using Elixir's regular Logger instead of
Appsignal.IntegrationLogger for two reasons. Firstly, since it's run in
a __before_compile__/1, the configuration isn't loaded yet, so
Appsignal.IntegrationLogger can't function. Also, it's not needed since
it's using the warn level, which would be printed anyway.