Set span names from event_handler
The current version of AppSignal's Phoenix instrumentation relies on
AppSignal's Plug instrumentation. The Plug instrumentation starts the
root span, to which the Phoenix instrumentation adds child spans for
events happening within the requests.

This patch removes the reliance on the Plug instrumentation by setting
the span name from
`Appsignal.Phoenix.EventHandler.phoenix_endpoint_stop/4`. That removes
the need for the Plug integration to create the root span.

This patch is a prototype for further testing. Besides proper testing,
it's missing a couple of important parts:

- [ ] The previous version of installing AppSignal for Phoenix (by using
  `Appsignal.Phoenix` in the host app's endpoint) is disabled, but the
  function does not yet throw a warning to tell users to remove it.
- [ ] There's no error handling for requests with `Plug.Conn`s that
  don't have a `:phoenix_action` and `:phoenix_controller` key.
- [ ] The current version sets the span's name, but doesn't include
  metadata like parameters. These have to be exposed in
  `Appsignal.Plug`

Also, in the future, this approach could also work for non-Phoenix Plug
apps, as the telemetry-based instrumentation is not specific to Phoenix
apps, but part of `Plug.Telemetry`. Currently, the fastest and most
useful way to ship some of this work seems to be to release is for
Phoenix apps first.