Add :telemetry-only Phoenix instrumentation (#56)
* Set span names from event_handler

The current version of AppSignal's Phoenix instrumentation relies on
AppSignal's Plug instrumentation. The Plug instrumentation starts the
root span, to which the Phoenix instrumentation adds child spans for
events happening within the requests.

This patch removes the reliance on the Plug instrumentation by setting
the span name from
`Appsignal.Phoenix.EventHandler.phoenix_endpoint_stop/4`. That removes
the need for the Plug integration to create the root span.

This patch is a prototype for further testing. Besides proper testing,
it's missing a couple of important parts:

- [ ] The previous version of installing AppSignal for Phoenix (by using
  `Appsignal.Phoenix` in the host app's endpoint) is disabled, but the
  function does not yet throw a warning to tell users to remove it.
- [ ] There's no error handling for requests with `Plug.Conn`s that
  don't have a `:phoenix_action` and `:phoenix_controller` key.
- [ ] The current version sets the span's name, but doesn't include
  metadata like parameters. These have to be exposed in
  `Appsignal.Plug`

Also, in the future, this approach could also work for non-Phoenix Plug
apps, as the telemetry-based instrumentation is not specific to Phoenix
apps, but part of `Plug.Telemetry`. Currently, the fastest and most
useful way to ship some of this work seems to be to release is for
Phoenix apps first.

* Move setname to stop event in EventHandlerTest

This fixes the tests for EventHandlerTest, but leaves the main tests
failing for now.

* Add EventHandler.set_name/1 to set names

Instead of matching the action and controller names in
phoenix_endpoint_stop/4, take the whole conn and pass it to the newly
added function.

* Handle errors through EventHandler

Instead of relying on Appsignal.Plug's wrapper block, the Phoenix
integration now listens for exception events reported direcly from
Phoenix.

* Set params through EventHandler

Params are set in the close span event, to make sure they're fetched
before they're set.

* Remove Appsignal.PhoenixTest

The tests in Appsignal.PhoenixTest test functionality that's moved the
Appsignal.EventHandler. This patch removes the leftover tests.

* Set sample data from event handler

Call Appsignal.Span.set_sample_data/3 to set the environment from
EventHandler.

* Move EventHandlerTest conn to helper function

Reuse the conn from EventHandlerTest.conn/0

* Add regression test to set sample data on error

EventHandler.add_error/4 calls out to .set_conn_data/2, which sets the
name and sample data.

* Switch to Appsignal.Plug.set_conn_data/1

Instead of reimplementing the metadata functions in the Phoenix library,
use the newly updated metadata functions in Appsignal.Plug.

To make this work, the implementations for Appsignal.Metadata in this
library need to be implemented, so they're delegated to the Any
implementation.

* Add deprecation warning for use Appsignal.Phoenix

> warning: use Appsignal.Phoenix is deprecated. Since AppSignal for
> Phoenix 2.3.0, Phoenix instrumentation is up automatically. The `use
> Appsignal.Phoenix` line is no longer needed and should be removed from
> your app's endpoint file.

* Depend on Appsignal for Plug 2.0.13

2.0.13 adds Metadata functions needed for the updated Phoenix event
  handler.

* Add changeset to describe telemetry-only install

"Add :telemetry-only Phoenix instrumentation to remove the need for
includes in Phoenix application endpoints"