Don’t run Appsignal.Phoenix.View tests on Elixir < 1.7 (#65)
* Remove compilers from mix.exs

Before, we had to explicitly add the :phoenix compiler to the list of
compilers. In Phoenix 1.7, this now throws a warning on compile.

* Remove Appsignal.ViewTest to fix the build

Phoenix 1.7 removes Phoenix.View, which Appsignal.View is based on. This
commit removes the tests for Appsignal.View to fix the build. These will
be reinstated in a future patch.

* Run Appsignal.Phoenix.View tests on Phoenix < 1.7

This reverts commit 3521d77c1604468944b9996ed60c01b66cc1b77b.

This commit reinstates the tests for Appsignal.Phoenix.View, but only
runs them when Phoenix.View is available (on versions before 1.7).