Add catch-all to prevent event handler crashes
As reported in
<https://github.com/appsignal/appsignal-elixir/issues/831>, there are
situations where the error handler is detached if it fails to extract
error metadata. Although we haven't found the cause of the issue yet,
this patch adds a catch-all with a warning that prints the received
data. This should give us insight into what's happening in this specific
case, but also adds a clause so we're sure we don't crash the handler in
the future.