Set action names from router_dispatch_stop
Instead of using endpoint_stop, use router_dispatch_stop to set the
conn data, including the action name. This should match plug routes
without a plug_route in the conn (like forwarded requests).

This is a proof of concept. In the definitive implementation, the name
extraction needs to be moved to Appsignal.Metadata.