Start root spans through router_dispatch events
Instead of using endpoint events to start and stop the root span, switch
to the router_dispatch events. They're earlier in the stack, and include
the router key, which we'll use when we can't deduce the action name
from the conn, like when Plug apps are nested in Phoenix apps.