Use router_dispatch events instead of endpoint events to start root spans (#67)
* Replace Appsignal.Plug.set_span_data/2

Because the telemetry integration extracts data from the telemetry
metadata, and not just from the passed conn, the implementation
doesn't match Phoenix anymore. To prepare for future changes, move the
set_span_data/2 function into Appsignal.Phoenix.EventHandler.

* Start root spans through router_dispatch events

Instead of using endpoint events to start and stop the root span, switch
to the router_dispatch events. They're earlier in the stack, and include
the router key, which we'll use when we can't deduce the action name
from the conn, like when Plug apps are nested in Phoenix apps.

* Revert to phoenix_endpoint category

Instead of switching to phoenix_router_dispatch as the category name,
stay on phoenix_endpoint for now, to revisit the event names later.