Use Appsignal.Utils.warning/1 instead of Logger.warn/1 (#70)
* Use Logger.warning/1 instead of Logger.warn/1

Drop support for elxir 1.9 and 1.10 so we can move from warn to warning
log level.

* Use Utils.warning/1 instead of Logger.warning/1

Update the minimal version of the appsignal dependenvy to 2.7.4, which
is the version that introduces the backwards compatible logging
functions in Appsignal.Utils. Then, use Appsignal.Utils.warning/1 in
Appsignal.Phoenix.View and Appsignal.Phoenix.EventHandler to support
Elixir 1.10 through 1.15.

* Regenerate CI matrix

Update the CI matrix based on the versions listed in
appsignal-elixir's versions.rb file. Note that this does *not* include
any of the Elixir release candidates, as they break the build.

We're awaiting a release of ssl_verify_fun that includes
https://github.com/deadtrickster/ssl_verify_fun.erl/pull/27 before we
can add back the new Elixir verions.

* Add changeset to explain changes for patch version

"Fix Logger deprecation warnings on Elixir 1.15"

* Update appsignal_plug to fix Elixir 1.15 warnings

AppSignal for Plug 2.0.14 uses the backwards compatible logger. Depend
on that version or higher to solve issues in Phoenix app that originate
in the Plug library.

---------

Co-authored-by: Mark Madsen <mm@idyll.io>