Allow a fun to be given
Because
This change allows a fun to be given to attach that can be executed in
the handle_event_start. This can be used to set custom data on the fly.

WIth this we can take the OpenTelementry span that's used for HoneyComb
and send a link to HoneyComb as custom data to AppSignal. This is
possible because adding a fun it's only executing the needed span when
executed.