Determine action name from ActionClauseErrors
When a Phoenix.ActionClauseError is detected, it will not be able to set
the controller and action name as the span (action) name. The connection
doesn't have the controller action on its object because there's a
function signature mismatch.

The controller action it expects to find can be fetched from the error
object. That's what this change does specifically for the
Phoenix.ActionClauseError.