Notify changelog repository of new releases
Automate updating our changelog on the appsignal.com/changelog website.

This workflow will notify the `workflow_dispatch` workflow in the
appsignal.com repo a new changelog entry needs to be made.
Related PR: https://github.com/appsignal/appsignal.com/pull/1790

This requires new releases to be made with annotated tags created with
the latest version of Mono.
Related PR: https://github.com/appsignal/mono/pull/75

Cross repository communication via the GitHub API is done with a
Personal Access Token (PAT), that expires every year. That's the longest
GitHub will allow it be active for. This token will need to updated for
all repositories we add such an action to. Let's configure a reminder
for that somewhere. GitHub will also notify the token owner of this
before it expires.

[skip changeset]
[skip review]