Handle :phoenix, :endpoint events in EventHandler
Previously, the EventHandler ignored endpoint events, mostly because
of difficulty determining the root span, as the endpoint and
router_dispatch spans close in reverse order. Because of the addition
of the root_span functions, this can now be solved, as span data can
be set on the root span directly.

This patch moves the span data to the phoenix_endpoint_stop function,
which also causes spans to start earlier in the request. That, in
turn, means that plugs can be instrumented without changes once again.

Closes #83.