Fix Dialyzer warning in EventHandler
Currently, running dialyzer over the project produces this warning:

    lib/appsignal_phoenix/event_handler.ex:48:unmatched_return
    The expression produces a value of type:

    nil | %Appsignal.Span{:pid => pid(), :reference => reference()}

    but this value is unmatched.

This is caused by a call to set_span_data/2 in the
phoenix_envpoint_stop/4 function in the EventHandler. Here, Dialyzer
expects the return value to be used, probably because it's either a
nil or a Span.

In reality, we don't care what the return value is, as we're setting
span data on the root span if it exists. If there's no root span, we
don't need to set any metadata because we're trying to close a span
that doesn't exist.

To fix the warning, this patch explicitly ignores the return value by
putting the value in a variable prefixed with an underscore.