Move CI to Github actions (#93)
* Move CI to Github actions

Because of difficulty updating the test matrix and getting
dependencies installed on Semaphore, move the CI to Github actions.

[skip changeset]

* Fix compilation warning in EventHandler

In the phoenix_router_dispatch_stop/4 function in the EventHandler, a
parameter is passed named metadata that's never used. This produces a
warning during compilation.

This patch prefixes the variable with an underscore to indicate is's
unused.

* Fix Dialyzer warning in EventHandler

Currently, running dialyzer over the project produces this warning:

    lib/appsignal_phoenix/event_handler.ex:48:unmatched_return
    The expression produces a value of type:

    nil | %Appsignal.Span{:pid => pid(), :reference => reference()}

    but this value is unmatched.

This is caused by a call to set_span_data/2 in the
phoenix_envpoint_stop/4 function in the EventHandler. Here, Dialyzer
expects the return value to be used, probably because it's either a
nil or a Span.

In reality, we don't care what the return value is, as we're setting
span data on the root span if it exists. If there's no root span, we
don't need to set any metadata because we're trying to close a span
that doesn't exist.

To fix the warning, this patch explicitly ignores the return value by
putting the value in a variable prefixed with an underscore.

* Remove semaphore directory and scripts

Remove the .semaphore directory, and remove all scripts that were used
to get semaphore to run the test suite. Those are no longer needed
after switching to Actions.