Set render metadata as tags
When an error happens during rendering, the backtrace of it is
limited to the Elixir process that is spawned for the rendering task,
lacking any context from the customer's own code. To ameliorate this
issue, add the rendering metadata as tags to the root span, providing
some additional context in the resulting error sample.