Instrument views and components separately
Before this change, traces representing method calls in live views
would be grouped with traces representing method calls in live
components, and represented as calls to those views, rather than
to the components.

That is, when `SomeLiveComponent.handle_event/3` was called, the
corresponding trace in AppSignal would not be named
`SomeLiveComponent#handle_event`, but `SomeLiveView#handle_event`,
where `SomeLiveView` is the view in which `SomeLiveComponent`
happened to be mounted for that specific invocation.

This also meant that, for methods such as `update`, which exist for
both views and components, the samples representing calls to the
method on the view would be grouped together with the samples
representing calls to the update method in any of the components that
were mounted in the view, and the resulting events would be identical.

This means that it is not possible for the customer to obtain
meaningful performance metrics for either the live view's update
method or for the components that are mounted within it, as the two
are mixed together.

This commit changes this behaviour, so that traces and events
representing calls to live components, and not to live views, are
given the `.live_component` event group, have the name of the
corresponding component in their event title and action name, and are
grouped under the `live_component` namespace.