Add event names to the action name
The `handle_event/3` callback on both views and components is usually
implemented using pattern matching to match different event names to
different function bodies.

This means it is valuable to discriminate between those events, as
they represent different code paths with different performance. These
event names should be low-cardinality, as dynamic data would be
provided in the params, not as part of the event name.