Do not `raise` when `CI` env is present
We introduced an internal environment variable to be able to test
AppSignal for Phoenix against different Phoenix versions. We
accidentally made it `raise` when the `CI` environment variable is
present, causing it to break in customers' CI.

This commit modifies the `mix.exs` so it does not `raise` under any
circumstances.